new/Consciousness_analysis/requirements.txt
# Core dependencies
openai>=1.0.0
python-dotenv>=1.0.0
flask>=3.0.0

# Retry logic and resilience
tenacity>=8.2.0

# Standard library enhancements
pathlib>=1.0.1

# Optional: For enhanced logging (if needed)
# colorlog>=6.7.0

# Optional: For data analysis (if needed)
# pandas>=2.0.0
# numpy>=1.24.0

# Development dependencies (optional)
# pytest>=7.4.0
# black>=23.0.0
# flake8>=6.0.0
# mypy>=1.5.0